hw/arm/stm32f405: Add preliminary flash interface emulation support
The flash interface (FLASHIF) peripheral provides a control interface
for the SoC embedded flash memory on the STM32F4xx series devices.

This commit adds preliminary support for the flash interface peripheral
emulation, in order to support proper emulation of the firmware images
that use the STM32Cube driver, which configures and validates the
FLASH_ACR register during system initialisation.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>